requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Data & ML
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3

# Technical Analysis
ta==0.11.0

# NLP & Sentiment
vaderSentiment==3.3.2

# News API
newsapi-python==0.2.7

# Database
aiosqlite==0.19.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
joblib==1.3.2

# CORS
fastapi-cors==0.0.6
